+++ tests/tests_pytorch/trainer/connectors/test_checkpoint_connector.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 import os
 from unittest import mock
+from unittest.mock import Mock
 
 import pytest
 import torch
@@ -21,6 +22,7 @@ from pytorch_lightning import Trainer
 from pytorch_lightning.callbacks import ModelCheckpoint
 from pytorch_lightning.demos.boring_classes import BoringModel
 from pytorch_lightning.trainer.states import TrainerFn
+from pytorch_lightning.utilities.migration.utils import _set_version
 
 
 def test_preloaded_checkpoint_lifecycle(tmpdir):
@@ -31,26 +33,27 @@ def test_preloaded_checkpoint_lifecycle(tmpdir):
 
     connector = trainer._checkpoint_connector
 
-    assert not connector.resume_checkpoint_path
+    assert not connector._ckpt_path
     assert not connector._loaded_checkpoint
 
     connector.resume_start()
-    assert not connector.resume_checkpoint_path
+    assert not connector._ckpt_path
     assert not connector._loaded_checkpoint
     connector.resume_end()
-    assert not connector.resume_checkpoint_path
+    assert not connector._ckpt_path
     assert not connector._loaded_checkpoint
 
     ckpt_path = trainer.checkpoint_callback.best_model_path
     trainer = Trainer(default_root_dir=tmpdir, max_steps=2)
     connector = trainer._checkpoint_connector
     connector.resume_start(ckpt_path)
-    assert connector.resume_checkpoint_path == ckpt_path
+    assert connector._ckpt_path == ckpt_path
     assert connector._loaded_checkpoint
     assert isinstance(connector._loaded_checkpoint, dict)
     trainer.state.fn = TrainerFn.FITTING
     connector.resume_end()
-    assert not connector.resume_checkpoint_path
+    # not cleared until next restoration, as the user might access it through `trainer.ckpt_path`
+    assert connector._ckpt_path == ckpt_path
     assert not connector._loaded_checkpoint
 
 
@@ -166,3 +169,54 @@ def test_loops_restore(tmpdir):
             if fn2 != fn:
                 trainer_loop2 = getattr(trainer, f"{fn2}_loop")
                 trainer_loop2.load_state_dict.assert_not_called()
+
+
+def test_stateful_trainer_ckpt_path_support(tmp_path):
+    """Tests support for the pattern used by NeMo's experiment manager."""
+    model = BoringModel()
+
+    # dummy ckpt data
+    ckpt_data = {"state_dict": model.state_dict(), "optimizer_states": {}, "lr_schedulers": {}}
+    _set_version(ckpt_data, "2.0.0")
+
+    # save a "checkpoint"
+    ckpt_path = tmp_path / "foo.ckpt"
+    torch.save(ckpt_data, ckpt_path)
+
+    # mock model checkpoint instance that has saved a last checkpoint
+    model_checkpoint = Mock(spec=ModelCheckpoint)
+    last_path = tmp_path / "last.ckpt"
+    torch.save(ckpt_data, last_path)
+    model_checkpoint._find_last_checkpoints.return_value = {last_path}
+
+    trainer = Trainer(default_root_dir=tmp_path, fast_dev_run=True, callbacks=model_checkpoint)
+
+    # set the ckpt path statefully
+    trainer.ckpt_path = ckpt_path
+    trainer.fit(model)
+    assert trainer.ckpt_path == ckpt_path  # not automatically cleaned
+    assert trainer._checkpoint_connector._user_managed
+
+    # now conflict with ckpt_path functionally
+    with pytest.warns(UserWarning, match="trainer.ckpt_path =.*but then you passed"):
+        trainer.fit(model, ckpt_path="last")
+    assert trainer.ckpt_path == last_path
+    assert not trainer._checkpoint_connector._user_managed
+
+    # mock model checkpoint instance that has saved a last checkpoint
+    best_path = tmp_path / "best.ckpt"
+    torch.save(ckpt_data, best_path)
+    model_checkpoint.best_model_path = best_path
+
+    # `trainer.test` will use this over "best" if statefully set
+    trainer.ckpt_path = ckpt_path
+    trainer.test()
+    assert trainer.ckpt_path == ckpt_path
+
+    # ckpt_path = "best" still works if it's reset
+    trainer.ckpt_path = None
+    # the state is cleared
+    assert trainer._checkpoint_connector._ckpt_path is None
+    assert not trainer._checkpoint_connector._user_managed
+    trainer.test()
+    assert trainer.ckpt_path == best_path